Updated bot.py, added template spreadsheet (PlayerStats.xlsx) and initial Docker-related files
-fixed /rolepreference. It now updates user’s RolePreference value in the DB correctly, and sends an embed message displaying updated values to the user so that they can confirm their changes.
-updated documentation to remove mention of gspread/Google oauth
-added requirements.txt

Changes from week prior to 11/15:

-added error handling to /clear so users know when there’s an error with “Player” and “Volunteer” roles being higher in the hierarchy than the bot role
Added to update_player_rank() to include code handling an API connection error which would result in rank displaying as “N/A” on /playerstats. However, this is mostly untested and may not function properly as of yet. If all else fails, the error seems to only happen shortly after the bot is launched, and typing /playerstats multiple times resolves the error.
/playerstats has been changed to /stats to make typing both it, and /players, more convenient
Removed unused “api_config” dictionary