Updated Dockerfile, .dockerignore, README, and bot.py
The bot now runs successfully in a container, and will appear online when the container is run (with Docker Desktop AMD64 on Windows 10). However, any functionality apart from this is untested when using Docker.

.dockerignore has been updated to ignore all database files and spreadsheet files except for the templates included with the bot.

Unused imports from the spring 2024 team's version of the bot have been removed. Most of these were tied to Google Sheets integration.

Bot.py now automatically creates the "Player" and "Volunteer" roles when it is run in a server that does not already have them. This is performed by the on_ready() function.

Documentation in README.md has been updated.