Updated README.md and bot.py
Corrected line about privileged intents in readme, and removed the unused line "intents.message_content = True" in bot.py (as our bot only needed it when we were briefly considering the addition of alternative command prefixes e.g. "!"). This is not a change to the bot code, but I also disabled both Message Content and Presence intents for our bot application in the Discord Developer portal when the aforementioned changes were made, as we were not using those intents.